abstract core newPsychicApp logic to isolated files
No meaningful code changes happening, just moving the code out to
isolated files
